+++ test/core_test_app/app/components/component_loader.rb
@@ -27,17 +27,6 @@ class ComponentLoader < Netzke::Base
   # this action is using netzkeLoadComponent "special" callback
   action :load_with_feedback
 
-  # this action is using generic endpoint callback
-  action :load_with_generic_callback
-
-  # this action is using generic endpoint callback with scope
-  action :load_with_generic_callback_and_scope
-
-  endpoint :do_nothing do |params|
-    # here be tumbleweed
-#    {}
-  end
-
   action :load_component
 
   action :load_in_window
@@ -48,6 +37,8 @@ class ComponentLoader < Netzke::Base
 
   action :load_with_params
 
+  action :config_only
+
   action :non_existing_component do |a|
     a.text = "Non-existing component"
   end
@@ -56,7 +47,7 @@ class ComponentLoader < Netzke::Base
 
   def configure(c)
     super
-    c.bbar = [:load_component, :load_in_window, :load_with_feedback, :load_window_with_simple_component, :load_composite, :load_with_params, :load_with_generic_callback, :load_with_generic_callback_and_scope, :non_existing_component, :inaccessible]
+    c.bbar = [:load_component, :load_in_window, :load_with_feedback, :load_window_with_simple_component, :load_composite, :load_with_params, :non_existing_component, :inaccessible, :config_only]
   end
 
   endpoint :deliver_component do |params, this|